split cd file
